txshift: switch to uint for workers
workaround missing size_t implementation of
po::value_semantic* create_value(T& val, const std::string& arg)
as it would conflict with when size_t == unsigned int even when
it would help for size_t == unsigned long

fixes https://issues.apache.org/jira/browse/QPID-6135

Contributed by Dan Horaz <dan@danny.cz>

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@1653631 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/tests/txshift.cpp
@@ -40,7 +40,7 @@ namespace tests {
 struct Args : public qpid::TestOptions
 {
     std::string workQueue;
-    size_t workers;
+    uint workers;
 
     Args() : workQueue("txshift-control"), workers(1)
     {
@@ -178,7 +178,7 @@ int main(int argc, char** argv)
             worker.run();
         } else {
             boost::ptr_vector<Worker> workers;
-            for (size_t i = 0; i < opts.workers; i++) {
+            for (uint i = 0; i < opts.workers; i++) {
                 workers.push_back(new Worker(connection, opts.workQueue));
             }
             std::for_each(workers.begin(), workers.end(), boost::bind(&Worker::start, _1));
